# Speed up non-dominated sort