--- a/optuna/study/_multi_objective.py
+++ b/optuna/study/_multi_objective.py
@@ -188,9 +188,7 @@ def _calculate_nondomination_rank(
 
     # It ensures that trials[j] will not dominate trials[i] for i < j.
     # np.unique does lexsort.
-    unique_lexsorted_loss_values, order_inv = np.unique(
-        loss_values, return_inverse=True, axis=0
-    )
+    unique_lexsorted_loss_values, order_inv = np.unique(loss_values, return_inverse=True, axis=0)
 
     n_unique = unique_lexsorted_loss_values.shape[0]
     # Clip n_below.