… tests/study_tests/test_multi_objective.py
… tests/test_multi_objective.py