<!-- Thank you for creating a pull request! In general, we merge your pull request after it gets two or more approvals. To proceed to the review process by the maintainers, please make sure that the PR meets the following conditions: (1) it passes all CI checks, and (2) it is neither in the draft nor WIP state. If you wish to discuss the PR in the draft state or need any other help, please mention the Optuna development team in the PR. -->

## Motivation
<!-- Describe your motivation why you will submit this PR. This is useful for reviewers to understand the context of PR. -->

This PR relates to:
- https://github.com/optuna/optuna/issues/5089

This PR is a follow-up of 
- https://github.com/optuna/optuna/pull/5160
- https://github.com/optuna/optuna/pull/5160#issuecomment-1970399510

I specifically addressed:
- https://github.com/optuna/optuna/pull/5160#discussion_r1491890749
- https://github.com/optuna/optuna/pull/5160#issuecomment-1958851517
- https://github.com/optuna/optuna/pull/5160/files/4bfe8714ead38f3a1e232846ba044d7f39d031b9..542f0111b7b2da1fc662d486f91bf9252a163d72#r1491887538


## Description of the changes
<!-- Describe the changes in this PR. -->

By this PR, the non-dominated sort speeds up:

For the verification, I copied the implementations from the repo:

<details>
<summary>Used Implementations</summary>

```python
from __future__ import annotations

from collections import defaultdict
import time

import numpy as np


def _is_pareto_front_nd(unique_lexsorted_loss_values: np.ndarray) -> np.ndarray:
    # NOTE(nabenabe0928): I tried the Kung's algorithm below, but it was not really quick.
    # https://github.com/optuna/optuna/pull/5302#issuecomment-1988665532
    loss_values = unique_lexsorted_loss_values.copy()
    n_trials = loss_values.shape[0]
    on_front = np.zeros(n_trials, dtype=bool)
    nondominated_indices = np.arange(n_trials)
    while len(loss_values):
        nondominated_and_not_top = np.any(loss_values < loss_values[0], axis=1)
        # NOTE: trials[j] cannot dominate trials[j] for i < j because of lexsort.
        # Therefore, nondominated_indices[0] is always non-dominated.
        on_front[nondominated_indices[0]] = True
        loss_values = loss_values[nondominated_and_not_top]
        nondominated_indices = nondominated_indices[nondominated_and_not_top]

    return on_front


def _is_pareto_front_2d(unique_lexsorted_loss_values: np.ndarray) -> np.ndarray:
    n_trials = unique_lexsorted_loss_values.shape[0]
    cummin_value1 = np.minimum.accumulate(unique_lexsorted_loss_values[:, 1])
    is_value1_min = cummin_value1 == unique_lexsorted_loss_values[:, 1]
    is_value1_new_min = cummin_value1[1:] < cummin_value1[:-1]
    on_front = np.ones(n_trials, dtype=bool)
    on_front[1:] = is_value1_min[1:] & is_value1_new_min
    return on_front


def _is_pareto_front(unique_lexsorted_loss_values: np.ndarray) -> np.ndarray:
    (n_trials, n_objectives) = unique_lexsorted_loss_values.shape
    if n_objectives == 1:
        return unique_lexsorted_loss_values[:, 0] == unique_lexsorted_loss_values[0]
    elif n_objectives == 2:
        return _is_pareto_front_2d(unique_lexsorted_loss_values)
    else:
        return _is_pareto_front_nd(unique_lexsorted_loss_values)


def nd_rank_this_pr(
    loss_values: np.ndarray, *, n_below: int | None = None
) -> np.ndarray:
    if n_below is not None and n_below <= 0:
        return np.zeros(len(loss_values), dtype=int)

    (n_trials, n_objectives) = loss_values.shape
    if n_objectives == 1:
        _, ranks = np.unique(loss_values[:, 0], return_inverse=True)
        return ranks
    else:
        # It ensures that trials[j] will not dominate trials[i] for i < j.
        # np.unique does lexsort.
        unique_lexsorted_loss_values, order_inv = np.unique(
            loss_values, return_inverse=True, axis=0
        )

    n_unique = unique_lexsorted_loss_values.shape[0]
    # Clip n_below.
    n_below = min(n_below or len(unique_lexsorted_loss_values), len(unique_lexsorted_loss_values))
    ranks = np.zeros(n_unique, dtype=int)
    rank = 0
    indices = np.arange(n_unique)
    while n_unique - indices.size < n_below:
        on_front = _is_pareto_front(unique_lexsorted_loss_values)
        ranks[indices[on_front]] = rank
        # Remove the recent Pareto solutions.
        indices = indices[~on_front]
        unique_lexsorted_loss_values = unique_lexsorted_loss_values[~on_front]
        rank += 1

    ranks[indices] = rank  # Rank worse than the top n_below is defined as the worst rank.
    return ranks[order_inv]


def nd_rank_original(
    objective_values: np.ndarray,
    *,
    n_below: int | None = None,
    base_rank: int = 0,
) -> tuple[np.ndarray, int]:
    if n_below is not None and n_below <= 0:
        return np.full(len(objective_values), -1), base_rank
    if objective_values.shape[-1] == 1:  # NOTE(nabenabe0928): I changed here for debugging.
        _, ranks = np.unique(objective_values[:, 0], return_inverse=True)
        return ranks, np.max(ranks, initial=0)
    # Normalize n_below.
    n_below = n_below or len(objective_values)
    n_below = min(n_below, len(objective_values))

    # The ndarray `domination_mat` is a boolean 2d matrix where
    # `domination_mat[i, j] == True` means that the j-th trial dominates the i-th trial in the
    # given multi objective minimization problem.
    domination_mat = np.all(
        objective_values[:, np.newaxis, :] >= objective_values[np.newaxis, :, :], axis=2
    ) & np.any(objective_values[:, np.newaxis, :] > objective_values[np.newaxis, :, :], axis=2)

    domination_list = np.nonzero(domination_mat)
    domination_map = defaultdict(list)
    for dominated_idx, dominating_idx in zip(*domination_list):
        domination_map[dominating_idx].append(dominated_idx)

    ranks = np.full(len(objective_values), -1)
    dominated_count = np.sum(domination_mat, axis=1)

    rank = base_rank - 1
    ranked_idx_num = 0
    while ranked_idx_num < n_below:
        # Find the non-dominated trials and assign the rank.
        (non_dominated_idxs,) = np.nonzero(dominated_count == 0)
        ranked_idx_num += len(non_dominated_idxs)
        rank += 1
        ranks[non_dominated_idxs] = rank

        # Update the dominated count.
        dominated_count[non_dominated_idxs] = -1
        for non_dominated_idx in non_dominated_idxs:
            dominated_count[domination_map[non_dominated_idx]] -= 1

    return ranks, rank


def original(
    objective_values: np.ndarray,
    *,
    penalty: np.ndarray | None = None,
    n_below: int | None = None,
) -> np.ndarray:
    if penalty is None:
        ranks, _ = nd_rank_original(objective_values, n_below=n_below)
        return ranks

    if len(penalty) != len(objective_values):
        raise ValueError(
            "The length of penalty and objective_values must be same, but got "
            "len(penalty)={} and len(objective_values)={}.".format(
                len(penalty), len(objective_values)
            )
        )
    nondomination_rank = np.full(len(objective_values), -1)
    is_penalty_nan = np.isnan(penalty)
    n_below = n_below or len(objective_values)

    # First, we calculate the domination rank for feasible trials.
    is_feasible = np.logical_and(~is_penalty_nan, penalty <= 0)
    ranks, bottom_rank = nd_rank_original(
        objective_values[is_feasible], n_below=n_below
    )
    # NOTE(nabenabe0928): I changed here for debugging.
    ranks[ranks == -1] = np.max(ranks, initial=-1) + 1
    # NOTE(nabenabe0928): I changed here for debugging.
    bottom_rank = np.max(ranks, initial=-1)
    nondomination_rank[is_feasible] += 1 + ranks
    n_below -= np.count_nonzero(is_feasible)

    # Second, we calculate the domination rank for infeasible trials.
    is_infeasible = np.logical_and(~is_penalty_nan, penalty > 0)
    num_infeasible_trials = np.count_nonzero(is_infeasible)
    if num_infeasible_trials > 0:
        _, ranks = np.unique(penalty[is_infeasible], return_inverse=True)
        ranks += 1
        nondomination_rank[is_infeasible] += 1 + bottom_rank + ranks
        bottom_rank += np.max(ranks)
        n_below -= num_infeasible_trials

    # Third, we calculate the domination rank for trials with no penalty information.
    ranks, _ = nd_rank_original(
        objective_values[is_penalty_nan], n_below=n_below, base_rank=bottom_rank + 1
    )
    nondomination_rank[is_penalty_nan] += 1 + ranks

    return nondomination_rank


def this_pr(
    loss_values: np.ndarray,
    *,
    penalty: np.ndarray | None = None,
    n_below: int | None = None,
) -> np.ndarray:
    if penalty is None:
        return nd_rank_this_pr(loss_values, n_below=n_below)

    if len(penalty) != len(loss_values):
        raise ValueError(
            "The length of penalty and loss_values must be same, but got "
            f"len(penalty)={len(penalty)} and len(loss_values)={len(loss_values)}."
        )

    nondomination_rank = np.full(len(loss_values), -1, dtype=int)
    is_penalty_nan = np.isnan(penalty)
    n_below = n_below or len(loss_values)

    # First, we calculate the domination rank for feasible trials.
    is_feasible = np.logical_and(~is_penalty_nan, penalty <= 0)
    nondomination_rank[is_feasible] = nd_rank_this_pr(
        loss_values[is_feasible], n_below=n_below
    )
    n_below -= np.count_nonzero(is_feasible)

    # Second, we calculate the domination rank for infeasible trials.
    top_rank_in_infeasible = np.max(nondomination_rank[is_feasible], initial=-1) + 1
    is_infeasible = np.logical_and(~is_penalty_nan, penalty > 0)
    n_infeasible = np.count_nonzero(is_infeasible)
    if n_infeasible > 0:
        _, ranks_in_infeas = np.unique(penalty[is_infeasible], return_inverse=True)
        nondomination_rank[is_infeasible] = ranks_in_infeas + top_rank_in_infeasible
        n_below -= n_infeasible

    # Third, we calculate the domination rank for trials with no penalty information.
    top_rank_in_penalty_nan = np.max(nondomination_rank[~is_penalty_nan], initial=-1) + 1
    ranks_in_penalty_nan = nd_rank_this_pr(
        loss_values[is_penalty_nan], n_below=n_below
    )
    nondomination_rank[is_penalty_nan] = ranks_in_penalty_nan + top_rank_in_penalty_nan
    assert np.all(nondomination_rank != -1), "All the rank must be updated."
    return nondomination_rank


def verify(
    seed: int, n_samples: int, n_below: int, n_objectives: int, consider_penalty: bool = True
) -> tuple[float, float]:
    # TODO(nabenabe0928): consider_penalty.
    print(f"Start with {seed=}, {n_samples=}, {n_below=}, {n_objectives=}.")
    rng = np.random.RandomState(seed)
    loss_values = rng.normal(size=(n_samples, n_objectives))
    penalty = 1 - loss_values[:, 0] if consider_penalty else None

    start = time.time()
    output_this_pr = this_pr(loss_values, penalty=penalty, n_below=n_below)
    runtime_this_pr = (time.time() - start) * 1000
    print(f"This PR took {runtime_this_pr:.3f} ms.")

    start = time.time()
    output_original = original(loss_values, penalty=penalty, n_below=n_below)
    runtime_original = (time.time() - start) * 1000
    print(f"Original implementation took {runtime_original:.3f} ms.")

    assert np.allclose(output_original, output_this_pr), (
        f"{n_samples=}, {n_below_div=}, {n_objectives=}, "
        f"{output_original=}, {output_this_pr=}, {penalty=}"
    )
    return runtime_this_pr, runtime_original
```
</details>

> [!IMPORTANT]
> Please check `# NOTE(nabenabe0928): I changed here for debugging.` as some parts were just a bug and some parts hindered the reproducibility.

<details>
<summary>Code to Generate the Table</summary>

```python
import pandas as pd
def add_row(results: dict, row: dict) -> None:
    for k, v in row.items():
        results[k].append(v)

n_seeds = 5
n_samples_set = [1000, 10000]
n_objectives_set = [2, 3]
results = {"version": [], "n_objectives": [], "n_samples": [], "runtime": [], "seed": []}
for seed in range(n_seeds):
    for n_samples in n_samples_set:
        for n_objectives in n_objectives_set:
            row = dict(seed=seed, n_samples=n_samples, n_objectives=n_objectives)
            rt_this_pr, rt_original = verify(
                seed, n_samples, n_samples, n_objectives, consider_penalty=False
            )
            row.update(version="original", runtime=rt_original)
            add_row(results, row)
            row.update(version="this_pr", runtime=rt_this_pr)
            add_row(results, row)

posthoc_results = {"n_objectives": [], "n_samples": [], "Original": [], "This PR": []}
df = pd.DataFrame(results)
for n_samples in n_samples_set:
    for n_objectives in n_objectives_set:
        row = dict(n_samples=n_samples, n_objectives=n_objectives)
        cond1 = df["n_samples"] == n_samples
        cond2 = df["n_objectives"] == n_objectives
        df_original = df[cond1 & cond2 & (df["version"] == "original")]
        m = df_original["runtime"].mean()
        ste = df_original["runtime"].std() / np.sqrt(n_seeds)
        row["Original"] = f"{m:.3e} $\pm$ {ste:.3e}"

        df_this_pr = df[cond1 & cond2 & (df["version"] == "this_pr")]
        m = df_this_pr["runtime"].mean()
        ste = df_this_pr["runtime"].std() / np.sqrt(n_seeds)
        row["This PR"] = f"{m:.3e} $\pm$ {ste:.3e}"

        add_row(posthoc_results, row)

print(pd.DataFrame(posthoc_results).to_markdown(index=False))
```

</details>

|   n_objectives |   n_samples | Original                | This PR                 |
|---------------:|------------:|:------------------------|:------------------------|
|              2 |        1000 | 8.497e+01 $\pm$ 1.346e+00 | 1.167e+00 $\pm$ 3.161e-02 |
|              3 |        1000 | 6.647e+01 $\pm$ 1.619e+00 | 8.594e+00 $\pm$ 3.196e-01 |
|              2 |       10000 | 1.024e+04 $\pm$ 2.006e+02 | 1.743e+01 $\pm$ 6.119e-01 |
|              3 |       10000 | 7.322e+03 $\pm$ 1.500e+02 | 3.592e+02 $\pm$ 7.861e+00 |

> [!NOTE]
> The unit of the runtimes in this table is milliseconds.