@nabenabe0928 2024-03-11
I also tried the Kung's algorithm, but this algorithm was not really quick in my benchmarking.

```python
import time

import numpy as np


def is_pareto_front_nd(unique_lexsorted_loss_values: np.ndarray) -> np.ndarray:
    loss_values = unique_lexsorted_loss_values.copy()
    n_trials = loss_values.shape[0]
    on_front = np.zeros(n_trials, dtype=bool)
    nondominated_indices = np.arange(n_trials)
    while len(loss_values):
        nondominated_and_not_top = np.any(loss_values < loss_values[0], axis=1)
        # NOTE: trials[j] cannot dominate trials[j] for i < j because of lexsort.
        # Therefore, nondominated_indices[0] is always non-dominated.
        on_front[nondominated_indices[0]] = True
        loss_values = loss_values[nondominated_and_not_top]
        nondominated_indices = nondominated_indices[nondominated_and_not_top]

    return on_front


def kung(unique_lexsorted_loss_values: np.ndarray, vectorization: bool = False) -> np.ndarray:
    if unique_lexsorted_loss_values.shape[0] == 1:
        return np.array([0])

    n_points = unique_lexsorted_loss_values.shape[0]
    mid = n_points // 2
    top_non_dominated_indices = kung(unique_lexsorted_loss_values[:mid], vectorization)
    bottom_non_dominated_indices = kung(unique_lexsorted_loss_values[mid:], vectorization)
    top_non_dominated_values = unique_lexsorted_loss_values[:mid][top_non_dominated_indices]
    bottom_non_dominated_values = unique_lexsorted_loss_values[mid:][bottom_non_dominated_indices]

    bottom_non_dominated_by_top_indices = np.arange(n_points - mid)[bottom_non_dominated_indices]
    if vectorization:
        non_dominated = np.all(
            np.any(
                bottom_non_dominated_values[:, np.newaxis] < top_non_dominated_values,
                axis=-1,
            ), axis=-1
        )
        bottom_non_dominated_by_top_indices = bottom_non_dominated_by_top_indices[non_dominated]
    else:
        for loss_values in top_non_dominated_values:
            non_dominated = np.any(bottom_non_dominated_values < loss_values, axis=-1)
            bottom_non_dominated_values = bottom_non_dominated_values[non_dominated]
            bottom_non_dominated_by_top_indices = bottom_non_dominated_by_top_indices[
                non_dominated
            ]

    pareto_indices = np.hstack([
        top_non_dominated_indices,
        mid + bottom_non_dominated_by_top_indices,
    ])
    return pareto_indices


if __name__ == "__main__":
    rng = np.random.RandomState(42)
    x = rng.normal(size=(10000, 3))
    y = np.unique(x, axis=0)
    start = time.time()
    pareto_indices = kung(y)
    print(f"Kung: {time.time() - start}")

    start = time.time()
    ans = is_pareto_front_nd(y)
    print(f"Original: {time.time() - start}")
    output = np.zeros(y.shape[0], dtype=bool)
    output[pareto_indices] = True
    assert np.all(output == ans), y[np.where(output != ans)[0]]

```